pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apme-engine"
version = "0.1.0"
description = "Ansible quality/modernization engine: ARI-based scanner + OPA rules"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "jsonpickle",
    "PyYAML",
    "ruamel.yaml",
    "filelock",
    "rapidfuzz",
    "joblib",
    "grpcio>=1.78.0",
    "grpcio-health-checking>=1.78.0",
    "protobuf>=6.31.1,<7",
    "httpx",
]

[project.optional-dependencies]
ai = [
    "abbenay-client @ https://github.com/redhat-developer/abbenay/releases/download/v2026.3.8-alpha/abbenay_client-2026.3.8a0-py3-none-any.whl#sha256=b4badd60cac70a12624febe02b791110f1bb1f7da872f10ff3992304496ae306",
]
proxy = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]
gateway = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio>=0.24",
    "ruff",
    "mypy",
    "pydoclint>=0.8.0",
    "types-protobuf",
    "types-PyYAML",
    "grpc-stubs",
    "grpcio-tools>=1.78.0",
    "pytest-playwright",
]

[tool.pytest.ini_options]
testpaths = ["tests", "src/apme_engine/validators/native/rules"]
pythonpath = ["src"]
addopts = "-v --tb=short -m 'not integration and not ui'"
markers = [
    "integration: marks tests as integration (may use network or external deps)",
    "ui: marks tests as browser/Playwright tests (require running UI + gateway stack)",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/apme_engine"]
omit = ["*/tests/*", "*/__pycache__/*", "*/rules/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true
fail_under = 50

[project.scripts]
apme-scan = "apme_engine.cli:main"
apme-primary = "apme_engine.daemon.primary_main:main"
apme-ansible-validator = "apme_engine.daemon.ansible_validator_main:main"
apme-native-validator = "apme_engine.daemon.native_validator_main:main"
apme-opa-validator = "apme_engine.daemon.opa_validator_main:main"
apme-gitleaks-validator = "apme_engine.daemon.gitleaks_validator_main:main"
galaxy-proxy = "galaxy_proxy.cli:main"
apme-gateway = "apme_gateway.main:main"

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["src/apme/v1/*_pb2*.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "D"]

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["typing.overload"]
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"src/apme_engine/engine/annotators/ansible.builtin/*.py" = ["E501"]

[tool.mypy]
strict = true
disallow_any_explicit = true
python_version = "3.10"
mypy_path = "src"
exclude = ["src/apme/v1/.*_pb2"]

# Vendored galaxy-proxy: parsed Galaxy YAML/JSON is inherently unstructured.
[[tool.mypy.overrides]]
module = ["galaxy_proxy", "galaxy_proxy.*"]
disallow_any_explicit = false

# Gateway DB layer: SQLAlchemy mapped columns use generic types.
[[tool.mypy.overrides]]
module = ["apme_gateway", "apme_gateway.*"]
disallow_any_explicit = false

# httpx stubs type Response.content as bytes; ignore is for ambiguous/alternate stubs.
[[tool.mypy.overrides]]
module = "galaxy_proxy.galaxy_client"
disable_error_code = ["unused-ignore"]

[[tool.mypy.overrides]]
module = [
    "pytest",
    "ruamel.*",
    "ansible.*",
    "jsonpickle",
    "filelock",
    "rapidfuzz.*",
    "joblib",
    "httpx",
    "yaml",
    "grpc",
    "grpc.*",
    "grpc_health",
    "grpc_health.*",
    "abbenay_grpc",
    "abbenay_grpc.*",
    "fastapi",
    "fastapi.*",
    "uvicorn",
    "uvicorn.*",
    "sqlalchemy",
    "sqlalchemy.*",
    "aiosqlite",
    "aiosqlite.*",
    "pydantic",
    "pydantic.*",
    "playwright",
    "playwright.*",
]
ignore_missing_imports = true

# Deprecation scraper uses untyped JSON dicts extensively.
[[tool.mypy.overrides]]
module = [
    "scrape_ansible_deprecations",
]
disallow_any_explicit = false

[[tool.mypy.overrides]]
module = [
    "apme.v1.ansible_pb2",
    "apme.v1.ansible_pb2_grpc",
    "apme.v1.common_pb2_grpc",
    "apme.v1.primary_pb2_grpc",
    "apme.v1.validate_pb2_grpc",
    "apme.v1.reporting_pb2_grpc",
]
ignore_errors = true

[tool.pydoclint]
style = "google"
arg-type-hints-in-docstring = false
arg-type-hints-in-signature = true
check-return-types = false
skip-checking-short-docstrings = false
check-class-attributes = true
allow-init-docstring = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"apme_engine.engine" = ["*.txt", "*.json"]
"apme_engine.engine.annotators" = ["ansible.builtin/*.py"]
"apme_engine.validators.opa" = ["bundle/*.rego", "bundle/*.json", "bundle/*.md"]
"apme_engine.data" = ["*.yml", "*.yaml"]
"galaxy_proxy" = ["py.typed"]
